+++ app/src/main/java/com/radlance/fooddelivery/domain/entity/Order.kt
@@ -2,5 +2,5 @@ package com.radlance.fooddelivery.domain.entity
 
 data class Order(
     val product: Product,
-    val amount: Double
+    val amount: Int
 )
